+++ tiledb/sm/query/query.h
@@ -284,6 +284,23 @@ class Query {
   Status get_est_result_size(
       const char* name, uint64_t* size_off, uint64_t* size_val);
 
+  /**
+   * Gets the estimated result size (in bytes) for the input fixed-sized,
+   * nullable attribute.
+   */
+  Status get_est_result_size_nullable(
+      const char* name, uint64_t* size_val, uint64_t* size_validity);
+
+  /**
+   * Gets the estimated result size (in bytes) for the input var-sized,
+   * nullable attribute.
+   */
+  Status get_est_result_size_nullable(
+      const char* name,
+      uint64_t* size_off,
+      uint64_t* size_val,
+      uint64_t* size_validity);
+
   /** Retrieves the number of written fragments. */
   Status get_written_fragment_num(uint32_t* num) const;